// src/app/interfaces/recipe/recipe.model.ts
export interface Recipe {
  uri: string;
  label: string;
  image: string;
  source: string;
  url: string;
  yield: number;
  dietLabels: string[];
  healthLabels: string[];
  cautions: string[];
  ingredientLines: string[];
  ingredients: Ingredient[];
  calories: number;
  totalWeight: number;
  totalTime: number;
  cuisineType: string[];
  mealType: string[];
  dishType: string[];
}

export interface Ingredient {
  text: string;
  food: string;
  weight: number;
  foodCategory: string;
  foodId: string;
  image: string;
  quantity: number;
}